refactor(sandbox): rename fallback to unsafe pid executor
