+++ docs/self-hosting/deployment-options/aws-ecs.mdx
@@ -20,9 +20,9 @@ This includes both:
 - `core.script.run_python` actions
 - Custom registry actions that execute code pulled from third-party sources
 
-AWS Fargate deployments have `nsjail` **disabled by default**. Do **not** run these workloads in production until you explicitly enable it.
+AWS Fargate deployments have `nsjail` **disabled by default**. In that mode, script execution falls back to `unsafe_pid_executor` PID namespace isolation (when available), which is less isolated than `nsjail`.
 
-Fargate does not provide the required permissions model for `nsjail`. For production workloads that execute untrusted code, we recommend Kubernetes (Helm chart), where `nsjail` is enabled by default.
+Fargate does not provide the required permissions model for `nsjail`. For production workloads that execute untrusted code, we still recommend Kubernetes (Helm chart), where `nsjail` is enabled by default.
 
 For comparison: Helm chart / Kubernetes deployments have `nsjail` enabled by default.
 </Warning>
