+++ tracecat/sandbox/__init__.py
@@ -9,10 +9,9 @@ It supports two execution modes:
    - Resource limits (memory, CPU, file size)
    - Requires privileged Docker mode or CAP_SYS_ADMIN
 
-2. **Safe executor** (when TRACECAT__DISABLE_NSJAIL=true, default):
-   - AST-based script validation with deny-by-default imports
-   - Runtime import hooks to block unauthorized modules
+2. **Unsafe PID executor** (when TRACECAT__DISABLE_NSJAIL=true, default):
    - Subprocess isolation with venv per dependency set
+   - PID namespace isolation when available
    - Works without privileged mode
 
 Both modes support:
@@ -28,21 +27,17 @@ from tracecat.sandbox.exceptions import (
     SandboxTimeoutError,
     SandboxValidationError,
 )
-from tracecat.sandbox.safe_executor import (
-    NETWORK_MODULES,
-    SAFE_STDLIB_MODULES,
-    SYSTEM_ACCESS_MODULES,
-    SafePythonExecutor,
-    ScriptValidator,
-)
 from tracecat.sandbox.safe_lambda import SafeLambdaValidator, build_safe_lambda
 from tracecat.sandbox.service import SandboxService, validate_run_python_script
 from tracecat.sandbox.types import ResourceLimits, SandboxConfig, SandboxResult
+from tracecat.sandbox.unsafe_pid_executor import (
+    UnsafePidExecutor,
+)
 
 __all__ = [
     # Services
     "SandboxService",
-    "SafePythonExecutor",
+    "UnsafePidExecutor",
     # Types
     "ResourceLimits",
     "SandboxConfig",
@@ -55,11 +50,7 @@ __all__ = [
     "PackageInstallError",
     # Validation
     "validate_run_python_script",
-    # Security validators and constants
+    # Security validators
     "SafeLambdaValidator",
-    "ScriptValidator",
     "build_safe_lambda",
-    "SAFE_STDLIB_MODULES",
-    "NETWORK_MODULES",
-    "SYSTEM_ACCESS_MODULES",
 ]
